Make staffline patch ids more unique.
We could run the pipeline first to collect a small number of patches for
labeling, then re-run it, extracting more patches, and want to combine
all patches. As long as the pipeline logic doesn't change in between,
this should ensure that the patch numbers uniquely identify the patch in
the page.