Support reading the run min length from a constant in the saved model.
This will allow setting this runtime parameter on a per-model basis, to
tweak the model performance. If the model is very sensitive to x
coordinate, we might need to decrease run_min_length so that a shorter
run of consecutive patches can result in a glyph. Otherwise, we might
need to increase run_min_length to decrease false positives.

In a follow-up, will add a command-line tool for writing this type of
constant to an existing saved model.